renamed to maws directory
