+++ maws/structure.py
